+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [UNRELEASED] - 0000-00-00
+### Fixed
+- Root password being lost after vm create
 
 ## [1.0.0] - 2016-01-31
 ### Added